Reversed text bug fixed in copy_coordinates_from_xl_to_rahshal()
Text prints in revers bug solution :
Before that only the first table was right and the rest were reversed.
I solved it by defining a global variable that keeps tack of the number of tables that is already modified.
If table_index == 0 means that we are still on the first table.
If it is bigger then 0 then we have passed the first table.
So to fix the bug it intentionally prints the values in reverse and the final product is as expected !
a.k.a backward logic